# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axon API",
    "description": "Read API for a personal Matrix state layer. Account-scoped resources nest under /v1/accounts/{account_id}; /v1/rooms is the cross-account aggregate.",
    "license": {
      "name": "LicenseRef-axon",
      "identifier": "LicenseRef-axon"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/accounts/{account_id}/rooms/{room_id}/avatar": {
      "put": {
        "tags": [
          "room-settings"
        ],
        "summary": "Set this room's avatar from an already-staged upload.",
        "operationId": "set_room_avatar",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoomAvatarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Avatar set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request, or the upload isn't an image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account, room, or upload not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "room-settings"
        ],
        "summary": "Clear this room's avatar.",
        "operationId": "remove_room_avatar",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account or room not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/accounts/{account_id}/rooms/{room_id}/name": {
      "put": {
        "tags": [
          "room-settings"
        ],
        "summary": "Set this room's name. An empty `name` clears it.",
        "operationId": "set_room_name",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoomNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Name set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account or room not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/accounts/{account_id}/rooms/{room_id}/send-media": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Send a staged media upload to a room.",
        "operationId": "send_media",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Media sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SendResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account, room, or upload not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/accounts/{account_id}/rooms/{room_id}/topic": {
      "put": {
        "tags": [
          "room-settings"
        ],
        "summary": "Set this room's topic. An empty `topic` clears it.",
        "operationId": "set_room_topic",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "description": "Axon account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "room_id",
            "in": "path",
            "description": "Matrix room id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoomTopicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Topic set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or revoked bearer token",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                },
                "description": "RFC 6750 bearer challenge: `Bearer` for a missing or malformed credential, `Bearer error=\"invalid_token\"` for an unknown or revoked token."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account or room not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream homeserver error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Account not reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_Value": {
        "type": "object",
        "description": "Success envelope: a 2xx body is always `{ \"data\": <T> }`.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {}
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "The error envelope: `{ \"error\": { \"code\", \"message\" } }`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "The error detail."
          }
        }
      },
...
      "SendMediaRequest": {
        "type": "object",
        "description": "Request body for sending a staged media upload into a room\n(`POST …/rooms/{room_id}/send-media`). The `upload_id` must refer to an\nunsent staged upload owned by the path account.",
        "required": [
          "upload_id"
        ],
        "properties": {
          "caption": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional media caption. When absent, Axon uses the staged filename as the\nMatrix event body."
          },
          "reply_to": {
            "type": [
              "string",
              "null"
            ],
            "description": "Send as a reply to this event id (`m.in_reply_to`). Optional."
          },
          "thread_root": {
            "type": [
              "string",
              "null"
            ],
            "description": "Send into this thread, identified by its root event id\n(`rel_type: m.thread`). Optional; when set without `reply_to`, the\nthread root is used as the reply fallback."
          },
          "upload_id": {
            "type": "string",
            "format": "uuid",
            "description": "Server-issued staged upload id returned by `POST …/media/uploads`."
          }
        }
      },
...
      "SetRoomAvatarRequest": {
        "type": "object",
        "description": "Request body for setting a room's avatar (`PUT …/rooms/{room_id}/avatar`;\nADR 0068 M19d). Takes an already-staged upload id, mirroring\n[`SendMediaRequest`]'s `upload_id` — Axon has no route that hands a\nclient an `mxc://` URI to reference directly, so avatar-set reuses the\nsame staged-upload flow as sending media.",
        "required": [
          "upload_id"
        ],
        "properties": {
          "upload_id": {
            "type": "string",
            "format": "uuid",
            "description": "Server-issued staged upload id returned by `POST …/media/uploads`."
          }
        }
      },
      "SetRoomNameRequest": {
        "type": "object",
        "description": "Request body for setting a room's name (`PUT …/rooms/{room_id}/name`;\nADR 0068 M19d). An empty `name` clears it — the SDK has no separate\n\"remove\" primitive for name/topic (unlike avatar).",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "SetRoomTopicRequest": {
        "type": "object",
        "description": "Request body for setting a room's topic (`PUT …/rooms/{room_id}/topic`;\nADR 0068 M19d). An empty `topic` clears it, same convention as\n[`SetRoomNameRequest`].",
        "required": [
          "topic"
        ],
        "properties": {
          "topic": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    }
  ],
...
}